Generalise CameraRecording with CameraRecordingStatus
CameraRecording is a wrapper for multiple CameraRecordingStatus
CameraRecordingStatus is the status for a single camera
Generalise mqtt code for server for catching camera recording statuses
